(PC-2113): Add get_offerer_count() to offerer repository
